requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
filelock==3.14.0
fsspec==2024.6.0
huggingface-hub==0.23.3
idna==3.7
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
safetensors==0.4.3
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
